refactor : LadderConnectOrder Interfcae 작업
